+++ package.json
@@ -18,10 +18,10 @@
     "@types/node": "^14.14.2",
     "jest": "^26.6.3",
     "jest-cli": "^26.6.3",
-    "obsidian": "https://github.com/obsidianmd/obsidian-api/tarball/master",
+    "obsidian": "^0.13.11",
     "rollup": "^2.33.1",
     "ts-jest": "^26.4.3",
     "tslib": "^2.0.3",
-    "typescript": "^4.0.3"
+    "typescript": "^4.1.0"
   }
 }
