+++ src/constant.ts
@@ -1,0 +1,2 @@
+
+export const defaultHeight = "300px"
