+++ src/types/aspect-ratio.ts
@@ -1,0 +1,2 @@
+
+export type AspectRatioType = `${number}/${number}` | 'none';
